Unset chromosome and location if the chromosome is unknown
Stops producing a transformed chromosome value of "UN".

This also switches to explicitly specifying the columns we want.  For
CSV output, when using a keygroup, the header is determined by the
fields present in the first record which match the group.  A missing
field in the first record could by chance completely change the output
now that we're deleting fields in transform.recs.