Remove the total_in_gene and total_at_location columns from integration_summary
total_in_gene was incorrectly double-counting (because it's applied
after the GROUP BY) and total_at_location is just misleading and not
super-useful.

For both of these concepts, I think it makes more sense to use
subqueries and normal aggregates.  For example, unique_sites in
summary_by_gene is a better substitute for total_in_gene, and easier to
understand how it's calculated.  total_at_location is more clearly
disclosed by aggregating over integration_summary and counting when
location is repeated.

We can develop these and other potential views further as we develop
more specs for useful reports.