NCI-RID: Correct coordinate transformation to depend on chromosome orientation
If we sequenced out of the HIV 3ʹ LTR on the reverse (−) strand of the
chromosome, then the first chromosomal base is already correct in the
zero-origin, interbase coordinate space.  Sequencing out of the HIV 5ʹ
LTR on the forward (+) strand of the chromosome is equivalent.

Similarly, sequencing out of the HIV 5ʹ LTR on the reverse (−) strand of
the chromosome is equivalent to sequencing out of the HIV 3ʹ LTR on the
forward (+) strand of the chromosome, and thus should be adjusted
accordingly too.

		     k−1
		 HIV  ↓
 F (+)  5ʹ---- 5ʹ~~~3ʹ ----3ʹ
 R (−)  3ʹ---- 3ʹ~~~5ʹ ----5ʹ
	      ↑
	      k

Thanks to Evan for pointing out my mistake!

Note that the form below doesn't happen, because a 5ʹ end can only
ligate to a 3ʹ end.

		 HIV
 F (+)  5ʹ---- 3ʹ~~~5ʹ ----3ʹ
 R (−)  3ʹ---- 5ʹ~~~3ʹ ----5ʹ