NCI-RID: Correct interpretation and use of gene_orientation from source data
This mistake was uncovered by Mike and Wenjie because they realized that
for a specific gene, only two sets of orientation tuples for
(chromosome, gene) are expected.

While investigating the issue, it became clear that I misinterpreted
gene_orientation as the orientation of the provirus in the gene, not the
orientation of the gene itself.  Discerning the proviral
orientation_in_gene involves looking at the orientation in the
chromosome of both the provirus and the gene.