List known issues from Trello on website
A little API funtimes to render a list of known issues with links to
Trello on the ISDB homepage. We should probably tweak the names on those
cards a bit so they'll be more illuminating without clicking through.
(Especially since that's not a public board and not everyone looking at
this has a Trello account, but the links will be useful to people with
access, and we could conceivably grant some access.)