Liftover: Filter out unsupported UCSC-style landmark names, for now
The liftOver tool is from UCSC, and it outputs their own naming scheme
for landmarks rather than RefSeq accessions.  It's possible to do the
translation, or to even start using UCSC-style names instead of
RefSeq accessions in the database, but it's more work¹ that will
have to come later.  It doesn't affect very much data right now.

¹ https://trello.com/c/Hj7QeUvy