Export a BED file suitable for loading into the UCSC Genome Browser
Eventually this should get rolled into the ISDB::Exporter framework
once exports have specific formats they support.

The idea of a simple, naive "interestingness" score is introduced where
lower values are more "interesting".  The idea is to bubble to the top
genes with many integrations in a small number of unique locations
across a large number of subjects.  Or to put it another way, the score
is trying to value "conservation" of integration sites.  (It's pretty
naive, though!)