Don't double-encode and double-decode cached JSON as UTF-8
The JSON encoder doesn't need to since the open pragma used at the top
of the script sets an output encoding on stdout.

Similarly, decode_json() already handles the decoding, so we want to
slurp the raw bytes, not decode it during the slurp.

Mea culpa for getting this wrong the first time with the PubMed info.