Identify the integration environment of the source data as in vivo or in vitro
In vivo meaning that the integrations happened in vivo, and were
observed ex vivo.  In vitro meaning the integrations happened in culture
or are from an existing cell line.

For example, I might say, "The _in vitro_ integration data presented in
the Sherrill-Mix, et al. study may be useful as a control for the _in
vivo_ integration data presented in the Wagner, et al. study."