load-integrations: Provide a --no-in-vitro flag
Skipping integrations which happened in vitro is useful when
constructing a database of just in vivo data.  The opposite is unlikely
to be as useful since in vitro data is usually going to be used as
control or model data, and comparing it within the same database to in
vivo data will be desirable.