Replace monthly archives with the concept of frozen versions
Frozen versions freeze the web pages as well as the exports, and put a
banner at the top of the main page identifying the page as frozen.

The web output directory is restructured to look like:

  index.html
  version.json
  doc/
    Manual.html
  exports/
    integration-summary.csv
    integration-summary.json
    integration-summary.metadata.json
    ...
  frozen/
    <name>/
      index.html
      version.json
      doc/
        Manual.html
      exports/
        integration-summary.csv
        integration-summary.json
        integration-summary.metadata.json
        ...

where the latest version is at the top level and frozen versions are
stored in individual directories under the top-level frozen/ directory.

Version metadata is stored in new version.json files produced by
bin/version-metadata.  Normally this is managed by bin/generate-website,
which by default generates the special "latest" version.  A new option
to generate-website, --freeze-as <name>, lets you create a frozen
version by giving a URL-safe name.  The scripts and page template know
that frozen versions live in the output directory under frozen/.  All
the freezing user has to know is to run generate-website with
--freeze-as.