Rough code for censoring duplicates from the Malderelli dataset
The Malderelli data contains sequences and IS determined from both the 5
prime and 3 prime LTRs.  Reported locations within ±6bp of each other in
the same subject, same orientation, and opposite LTRs are assumed to
have some unknown level of overlap.  To resolve this, we identify such
sites, clump them together, and keep only the site observed the most
frequently.  This is a conservative estimate of the true data.

Doing so is a bit tricky, and for the purposes of turning this around
quick for our consumers, is a manually run process.  Eventually it
should be folded into the regular data prep pipeline for this source and
not rely on a two-stage process during which the data is first loaded
into the ISDB and then the duplicates deleted.