Transform Sherrill-Mix et al 2013 for ISDB
The Sherrill-Mix data is itself an aggregation of data sets from three
different previous studies and new experiments the author group did, and
that information is useful to us since there's a distinction between
Jurkat and primary cells being infected in vitro that we are interested
in. So we're excluding that source when loading NCI-RID and loading it
directly here. In some ways it would be better to go even further back
and get the IS sequences directly from the source papers, but as
Sherill-Mix et al have done that for us already we'll rely on their
curation efforts.