Fix a double-encoding bug in the export metadata and JSON exporter
Both the JSON encoder object and the file handles from which we
read/write said JSON data were marked UTF-8.  This causes a
double-encoding (and double-decoding).  Since encodings should be pushed
out to the edges of the code, leave the file handles marked as UTF-8 and
let the JSON encoder always deal in characters not bytes.