Rework export formatters so they manage their own filehandles
The exporter provides a location to the formatter now, instead of an
opaque filehandle, and so the exporter no longer has to keep filehandle
state around.  This pushes filehandles closer to where they matter.

The types provided by Types::LoadableClass ensure that the classes are
loaded, so we no longer need to load them ourselves.