Excel export: Produce files with stable digests across equivalent runs
Stable digests are important because they’re the basis for change
detection and overwriting of previous exports in bin/export.  That
conditional overwriting is very useful.  It means the timestamps in the
export metadata, and thus also on the web, reflect the last time the
file actually changed, not merely the last time it was generated
(nightly, in our case).

I isolated two two small sources of change between otherwise equivalent
export runs: the Excel workbook timestamp and the zipfile member
timestamps.

I eliminate the former by using the workbook’s API to explicitly set a
meaningless but static creation time.  This timestamp shows up in
Excel’s File → Properties dialog under the Statistics tab, but no one
should actually care about it ever.  Besides, the on-disk file itself
will still have a separate timestamp kept by the filesystem and
operating system, and those are shown in that same dialog.

The latter is eliminated by re-opening the .xlsx file with Archive::Zip
after creation and manually setting the timestamps for each member using
the provided API.  While this does require knowing that .xlsx files are
zipfiles, it doesn’t require knowing anything further about their
contents and thus isn’t much of an encapsulation concern.

Both sets of timestamps are set to the same date:

  1987-08-19 12:00:00 US/Eastern

I choose this because it’s a) after Archive::Zip’s limitation of dates
later than 1980, which precludes the traditional Unix epoch, and b) is
my birthday.