Sunshine, et al. (JVI 2016) in vitro data
Data provided by Rory Kirchner in personal communication.  The "HIV
orientation table" slide was provided by Sara Sunshine in personal
communication.

A further step to-come is deduping the evidence which is plausibly from
opposite ends of the same provirus, similar to what's done
after-the-fact for the Malderelli data in NCI-RID.