Exporter producing one GFF3 file per gene with IS
These GFF files use positions relative to the gene start, but expect the
landmark sequence to be in the direction of the chromosome.  This
matches up with a chromosome sequence fragment downloaded from GenBank
for a particular gene, such as those imported by Geneious when searching
NCBI Gene.  An example such fragment is:

  http://www.ncbi.nlm.nih.gov/nuccore/NC_000019.10?report=genbank&from=2511219&to=2702748

for the GNG7 gene.  One nice thing about such sequence fragments is that
they also include other overlapping features in the same region.