Wagner-2014-Science: Correct location of reverse-oriented proviruses
Strand wasn't properly accounted for when processing the Wagner, et al.
spreadsheet to extract the integration sites.  The larger coordinate is
the correct site when the sequence was from the reverse strand.  My
incorrect assumption was that sequences had been reverse complemented so
they were all in the same orientation regardless of which strand was
originally sequenced.

This mistake led to the locations of reverse-oriented proviruses being
dozens to hundreds of bases off (i.e. the length of the human sequence
fragment).  Sequences were not reverse complemented either.  In most
cases, this shouldn't change the genes overlapping the sites.

Caught by Evan.  ++!