Cache stats used by the web templates
The basic idea is a stats.json file which gets provided to the templates
for rendering.  The file is produced by bin/cache-stats, a simple
script.

Most of the queries require scanning the entire "integration" table to
calculate counts, fields used, etc.  This takes a while, and is not
always desirable to do during every website generation.  By decoupling
fetching the query results from rendering the templates, we can cache
the results and only update them when desired.