Don't scale score in BED file
Just use the raw multiplicity.  UCSC's Genome Browser expects the score
to be between 0 and 1000 for use by a color scale, but we explicitly
turn that off in the track header with useScore=0.  Raw multiplicity is
more useful too when processing the BED file as data instead of
browsing it.