Create three standard roles as part of database creation
The three standard roles isdb_admin, isdb_rw, and isdb_ro are created
and granted appropriate privileges.  While usable directly, its
recommended that these roles are used as groups for other, site-specific
user roles.  Keeping the standard role names static makes the tooling
and documentation easier to write and reason about.  create-database is
now suitable for both development and production usage.

This also fixes a bug with the REVOKE … FROM PUBLIC statement, which was
quietly doing nothing when the public schema was still owned by the
postgres user (the default on Linux even when the rest of the database
was owned by isdb_admin).