Filter out dual-LTR evidence only within samples, not across
The purpose of collapsing the evidence is to not double-count the same
provirus in the same biologic sample.  It can't be the same provirus if
the evidence isn't from the same sample (although it _may_ be a clonal
virus, but those are collapsed into multiplicity later).

By providing an additional source-specific --sample-key field, the
collapser/clumper can avoid collapsing incorrectly across samples.

This revives a few records from the Maldarelli data.