Incorporate an interactive "Top N genes" visualization into the website
This visualization started out as a standalone manuscript figure
prepared by me for Jim.  The interactive version is to-be-published on
our website alongside a static version for the print/PDF copy.

Further discussion lead to Jim/me wanting to include this visualization
by default in ISDB.  It's pretty self-contained, making it easy to
import.  Some changes to the spec were made to accommodate input data
which is unknown in advance, unlike the analysis data we plan to
publish.