Jitter the x-value of points in mini "top genes" plot
The number of subjects is an inherently stepped value which will tend to
have multiple genes per level, leading to overplotting of points.
Jittering is a common solution to problematic overplotting, and so we do
so by up to 0.25.  The associated data table makes the real values
clear.  The x scale's domain remains the unjittered data so that it
doesn't shift around from page load to page load.  Clamping the scale
ensures that the jitter doesn't make points go, quite literally, off the
charts.

I had to calculate jitter_direction in a separate Vega expression
because of a bug with if() which produced initially baffling behaviour.
See https://github.com/vega/vega-expression/pull/10 for the bug report
and patch.

Jittering was suggested by Evan during review, and for that 👍.