load-genes: Support a --reload flag to drop existing data first
Similar to the --reload flag for load-source.  The flag is useful for
load-genes because I've subset the GFF3 annotation file to include only
transcribed genes and now want to load it into a copied database,
replacing existing gene data.  While I could do that outside of
load-genes in psql, it's nice to have it all in the same transaction in
case stuff goes wrong.