Fix an off-by-one bug when intersecting IS with genes
For example, given the diagram below, an IS at position 2 (or 6)
shouldn't be considered within the gene's bounds of [2,6], because
although we use interbase coordinates, the gene is defined by the
bases it includes.  Put another way, IS at position 2 (or 6) is between
a genic base and a non-genic base instead of between two genic bases.

        →       ←       → ← gene start and end
     A T·C G C G A T     ·  IS proviral location
    0 1 2 3 4 5 6 7 8

The BETWEEN x AND y SQL operator does an _inclusive_ check of the
bounds, so we need to make it exclusive by shortening the range inwards.

I noticed this error when comparing a gene intersection dataset produced
by ISDB (from HIRIS) with a supposedly equivalent gene intersection
dataset generated in R.  The comparison was intended as a QC check of
the R code, but what it turned up was a bug in ISDB.

In order to not dramatically increase, by an order of magnitude, the
execution time of the updated queries, a new index is added on the
specific expressions used by the view and query.