Update documentation tooling to expect Pandoc 2
Pandoc 2 was a major release which broke backcompat in some ways.  The
most noticeable was that --smart was removed and replaced with the
[+-]smart format extension syntax.  A small HTML class rename affected
us too.  I audited for other issues, but didn't find any.

While we could do version detection and support both Pandoc 1 and 2, I'm
about to add a commit providing a statically-linked pandoc similar to
what we do for liftover, thus obviating the need to support version 1.