Rewrite intra-documentation links from .md → .html automatically
Using a small Lua filter during processing with Pandoc.  Evan wrote the
basis for this filter in another repo, and I adapted it for our use
here, including adding support for anchors (#…).

Though a minor thing, it's nice to not have to link to the .html
versions since those files don't exist in the repository.  One upshot,
for example, is that links now work when Github renders the Markdown.