Top genes viz: Load data via <script> JS files instead of JSON/CSV over HTTP
This change is in order to support the use case of browsing the website
locally, directly from the filesystem.  While loading data directly over
HTTP is a bit cleaner and a better example of using our exports in most
cases, we can't do HTTP requests without a web server.

The Vega specs are easy to switch to embedded JS instead of JSON, and
I've just done that.  The gene summary data exports are a little
trickier, and CSV is a no-go.  Instead, I've added a JS export type
which wraps the JSON export in some boilerplate to provide the data via
a "well-known" global object, ISDB.Exports.  Since the JS exports might
be more generally useful, I've not gone to the trouble of hiding the new
JS exports from the data downloads page or selectively running them.

I think these are reasonable trade-offs for the end result, which is a
static website that doesn't require a web server!