Standardize command line snippets in installation documentation
Some of them were code blocks, others were inline code snippets, which I
think made them harder to spot as "now run this" prompts.  For example,
it was easy to miss the "make deps" step when it was inline.  This
makes everything a big ol' code block, which makes it pretty easy to see
the commands to run.  It also standardizes on no leading prompt for the
create-database examples.