Add a "Getting started" doc wrapping up a more complete initial experience
This sort of thing was suggested during manuscript review, and it's a
good idea.  The idea is to move people from installation and setup to
loading example data and finally generating and viewing the example
website.  It ends by giving them next steps for applying what they've
done to their own data.