Getting started: Remove two examples of loading sources
In his suggested edits, Evan removed the first example and kept the for
loop, noting:

    My concern was that in a doc like this, intended to be run straight
    through, having two versions of the same step would be a distraction.

I thought that was a good concern and cause for change but also
preferred to keep the spelled-out, non-loop version.  The reasoning is
that the non-loop example is simpler and more straightforward, and
people can more easily make the logical leap about how to load a single
source.  Folks who know how to write loops or run things in parallel can
do that themselves without guidance.